+++ src/core/index.ts
@@ -2,6 +2,6 @@
 // https://github.com/thomas-crane/tsai
 export * from './client';
 export * from './lib-info';
-export * from './plugin-manager';
-export * from './plugin-module';
+export * from './library-manager';
+export * from './library-module';
 export * from './resource-manager';