+++ src/core/library-module.ts
@@ -1,0 +1,3 @@
+export * from './../decorators/hook-packet';
+export * from './../decorators/library';
+export * from './../services/logger';