+++ src/decorators/library.ts
@@ -1,5 +1,5 @@
 import 'reflect-metadata';
-import { PluginManager } from './../core/plugin-manager';
+import { LibraryManager } from './../core/library-manager';
 import { LibraryInfo } from './../models/plugin-info';
 
 /**
@@ -10,7 +10,7 @@ export function Library(libInfo: LibraryInfo): ClassDecorator {
   return (target: any) => {
     const params = Reflect.getMetadata('design:paramtypes', target) || [];
     const dependencies = params.map((type: any) => type.name);
-    PluginManager.loadLibrary({
+    LibraryManager.loadLibrary({
       info: libInfo,
       target,
       dependencies