+++ src/models/index.ts
@@ -6,6 +6,7 @@ export * from './classes';
 export * from './condition-effect';
 export * from './enemy';
 export * from './environment';
+export * from './failure-code';
 export * from './guildrank';
 export * from './mapinfo';
 export * from './move-records';