+++ src/stdlib/index.ts
@@ -1,0 +1,6 @@
+// this index file is manually maintained
+// so it can be ignored by tsai. If it is
+// not ignored, the entire stdlib is loaded
+// at runtime, which is not desirable.
+export * from './object-tracker';
+export * from './player-tracker';